requirements.txt
BeautifulSoup==3.2.1
Django==1.4.3
Pillow==2.6.1
South==0.7.6
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-cache-utils==0.7.2
django-compressor==1.1
django-debug-toolbar==1.2.2
django-dotenv==1.2
django-extensions==1.4.8
django-getenv==1.3.1
django-robots==0.9.1
feedparser==5.1.3
gnureadline==6.3.3
guess-language==0.2
gunicorn==19.1.1
ipython==2.3.1
isodate==0.4.9
lxml==3.4.1
mock==1.0.1
psycopg2==2.5.4
python-amazon-product-api==0.2.2
python-memcached==1.48
pytz==2014.10
readline==6.2.4.1
requests==1.1.0
six==1.8.0
sqlparse==0.1.14
wsgiref==0.1.2
-e .
